+++ src/components/charts/index.ts
@@ -1,1 +1,2 @@
 export { LaunchesByYear } from './LaunchesByYear';
+export { SuccessRate } from './SuccessRate';